Network.Info

Documentation

getNetworkInterfaces :: IO [NetworkInterface]

Gets the address information for each of the network interfaces on the local computer.

data NetworkInterface

Describes the basic configuration of a network interface. This definition is currently limited to just one address per family.

Constructors

NetworkInterface

Fields

name :: String

Interface name (e.g. "eth0", "lo", "Local Area Connection")

ipv4 :: IPv4

IPv4 address

ipv6 :: IPv6

IPv6 address

mac :: MAC

MAC address

data IPv4

Represents an IPv4 address (e.g. 172.23.21.1, 127.0.0.1)

Constructors

IPv4 !Word32

data IPv6

Represents an IPv6 address (e.g. 2001:db8:85a3::8a2e:370:7334, ::1)

Constructors

IPv6 !Word32 !Word32 !Word32 !Word32

data MAC

Represents a MAC address (e.g. 01:23:45:67:89:ab)

Constructors

MAC !Word8 !Word8 !Word8 !Word8 !Word8 !Word8